use rabbitmq-plugin from symlinked directory
Currently, the rabbitmq management feature plugin is enabled
by calling a version that lives in the rabbitmq_server-2.7.1
directory.

This is a little brittle and couples the module to a very specific
version of rabbitmq. This patch updates the path to use the more
generic symlink: /usr/lib/rabbitmq/bin/rabbitmq-plugins